Refactor Watchlist and Up Next logic into Continue Watching and Watchlist Prefs
This change refactors the watchlist and "up next" domains to better reflect their responsibilities. Specifically, it separates user preference persistence from the background synchronization and data observation logic.

* **Data Layer Refactoring**:
    * Rename `data:watchlist` to `data:watchlist-prefs`, focusing specifically on list style and sort option persistence.
    * Remove `WatchlistDao` and `WatchlistRepository` in favor of more specialized data access objects.
    * Add `WatchlistPrefsRepository` and its implementation to manage UI state preferences.

* **Domain Layer Refactoring**:
    * Create `domain:continue-watching` to house logic previously located in `domain:watchlist` and `domain:upnext`.
    * Rename `WatchlistSyncInteractor` to `SyncContinueWatchingInteractor` and update it to use `SyncActivityInteractor`.
    * Rename `WatchlistTasksInitializer` to `ContinueWatchingTasksInitializer` and `WatchlistSyncWorker` to `ContinueWatchingSyncWorker`.
    * Move and rename `SyncWatchedShowInteractor` to `SyncShowMetadataInteractor` within `domain:showdetails`. It now includes syncing watch providers.
    * Migrate and rename models and interactors for `UpNext` and `WatchlistSections` to the `continuewatching` package.

* **Build & Configuration**:
    * Update `settings.gradle.kts` to reflect the new module structure: added `watchlist-prefs`, `continue-watching`, and `sync-activity`; removed `upnext` and `watchlist`.
    * Update `build.gradle.kts` files across affected modules to reflect dependency changes.

* **Testing**:
    * Migrate and update unit tests for the renamed interactors and repositories, ensuring metadata synchronization covers show details, seasons, episodes, and watch providers.
