+++ data/watchlist-prefs/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/watchlistprefs/api/model/WatchlistSortOption.kt
@@ -1,0 +1,10 @@
+package com.thomaskioko.tvmaniac.watchlistprefs.api.model
+
+public enum class WatchlistSortOption {
+    ADDED_DESC,
+    ADDED_ASC,
+    RELEASED_DESC,
+    RELEASED_ASC,
+    TITLE_ASC,
+    TITLE_DESC,
+}
