+++ data/watchlist-prefs/implementation/build.gradle.kts
@@ -1,0 +1,25 @@
+plugins {
+    alias(libs.plugins.app.kmp)
+}
+
+scaffold {
+    useMetro()
+}
+
+kotlin {
+    sourceSets {
+        commonMain {
+            dependencies {
+                api(libs.coroutines.core)
+                api(projects.data.datastore.api)
+                api(projects.data.watchlistPrefs.api)
+            }
+        }
+
+        commonTest {
+            dependencies {
+                implementation(libs.bundles.unittest)
+            }
+        }
+    }
+}
