--- data/watchlist/api/build.gradle.kts
@@ -1,14 +1,0 @@
-plugins {
-    alias(libs.plugins.app.kmp)
-}
-
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-                api(projects.data.database.sqldelight)
-                api(libs.coroutines.core)
-            }
-        }
-    }
-}
