--- data/watchlist/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/shows/api/model/WatchlistSortOption.kt
@@ -1,10 +1,0 @@
-package com.thomaskioko.tvmaniac.shows.api.model
-
-public enum class WatchlistSortOption {
-    ADDED_DESC,
-    ADDED_ASC,
-    RELEASED_DESC,
-    RELEASED_ASC,
-    TITLE_ASC,
-    TITLE_DESC,
-}
