--- data/watchlist/implementation/build.gradle.kts
@@ -1,29 +1,0 @@
-plugins {
-    alias(libs.plugins.app.kmp)
-}
-
-scaffold {
-    useMetro()
-}
-
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-                api(libs.coroutines.core)
-                api(projects.core.base)
-                api(projects.data.database.sqldelight)
-                api(projects.data.datastore.api)
-                api(projects.data.watchlist.api)
-
-                implementation(libs.sqldelight.extensions)
-            }
-        }
-
-        commonTest {
-            dependencies {
-                implementation(libs.bundles.unittest)
-            }
-        }
-    }
-}
