--- data/watchlist/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/watchlist/implementation/DefaultWatchlistRepository.kt
@@ -1,51 +1,0 @@
-package com.thomaskioko.tvmaniac.watchlist.implementation
-
-import com.thomaskioko.tvmaniac.datastore.api.DatastoreRepository
-import com.thomaskioko.tvmaniac.datastore.api.ListStyle
-import com.thomaskioko.tvmaniac.db.FollowedShows
-import com.thomaskioko.tvmaniac.db.SearchFollowedShows
-import com.thomaskioko.tvmaniac.shows.api.WatchlistDao
-import com.thomaskioko.tvmaniac.shows.api.WatchlistRepository
-import com.thomaskioko.tvmaniac.shows.api.model.WatchlistSortOption
-import dev.zacsweers.metro.AppScope
-import dev.zacsweers.metro.ContributesBinding
-import dev.zacsweers.metro.SingleIn
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.map
-
-@SingleIn(AppScope::class)
-@ContributesBinding(AppScope::class)
-public class DefaultWatchlistRepository(
-    private val watchlistDao: WatchlistDao,
-    private val datastoreRepository: DatastoreRepository,
-) : WatchlistRepository {
-
-    override fun observeWatchlist(): Flow<List<FollowedShows>> =
-        watchlistDao.observeShowsInWatchlist().distinctUntilChanged()
-
-    override fun searchWatchlistByQuery(query: String): Flow<List<SearchFollowedShows>> {
-        return watchlistDao.observeWatchlistByQuery(query)
-    }
-
-    override fun observeListStyle(): Flow<Boolean> {
-        return datastoreRepository.observeListStyle().map { listStyle ->
-            listStyle == ListStyle.GRID
-        }
-    }
-
-    override suspend fun saveListStyle(isGridMode: Boolean) {
-        val listStyle = if (isGridMode) ListStyle.GRID else ListStyle.LIST
-        datastoreRepository.saveListStyle(listStyle)
-    }
-
-    override fun observeSortOption(): Flow<WatchlistSortOption> =
-        datastoreRepository.observeWatchlistSortOption().map { name ->
-            WatchlistSortOption.entries.find { it.name == name }
-                ?: WatchlistSortOption.ADDED_DESC
-        }
-
-    override suspend fun saveSortOption(sortOption: WatchlistSortOption) {
-        datastoreRepository.saveWatchlistSortOption(sortOption.name)
-    }
-}
