--- data/watchlist/testing/build.gradle.kts
@@ -1,15 +1,0 @@
-plugins {
-    alias(libs.plugins.app.kmp)
-}
-
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-                api(projects.data.watchlist.api)
-                api(projects.data.database.sqldelight)
-                api(libs.coroutines.core)
-            }
-        }
-    }
-}
