+++ domain/continue-watching/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/continuewatching/model/EpisodeBadge.kt
@@ -1,0 +1,7 @@
+package com.thomaskioko.tvmaniac.domain.continuewatching.model
+
+public enum class EpisodeBadge {
+    PREMIERE,
+    NEW,
+    NONE,
+}
