+++ domain/continue-watching/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/continuewatching/model/UpNextResult.kt
@@ -1,0 +1,8 @@
+package com.thomaskioko.tvmaniac.domain.continuewatching.model
+
+import com.thomaskioko.tvmaniac.upnext.api.model.UpNextEpisode
+
+public data class UpNextResult(
+    val sortOption: UpNextSortOption,
+    val episodes: List<UpNextEpisode>,
+)
