+++ domain/continue-watching/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/continuewatching/model/UpNextSortOption.kt
@@ -1,0 +1,6 @@
+package com.thomaskioko.tvmaniac.domain.continuewatching.model
+
+public enum class UpNextSortOption {
+    LAST_WATCHED,
+    AIR_DATE,
+}
