--- domain/upnext/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/upnext/model/UpNextResult.kt
@@ -1,8 +1,0 @@
-package com.thomaskioko.tvmaniac.domain.upnext.model
-
-import com.thomaskioko.tvmaniac.upnext.api.model.UpNextEpisode
-
-public data class UpNextResult(
-    val sortOption: UpNextSortOption,
-    val episodes: List<UpNextEpisode>,
-)
