--- domain/upnext/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/upnext/model/UpNextSortOption.kt
@@ -1,6 +1,0 @@
-package com.thomaskioko.tvmaniac.domain.upnext.model
-
-public enum class UpNextSortOption {
-    LAST_WATCHED,
-    AIR_DATE,
-}
