--- domain/watchlist/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/watchlist/di/WatchlistTasksInitializerBindingContainer.kt
@@ -1,23 +1,0 @@
-package com.thomaskioko.tvmaniac.domain.watchlist.di
-
-import com.thomaskioko.tvmaniac.core.base.AsyncInitializers
-import com.thomaskioko.tvmaniac.core.base.Initializer
-import com.thomaskioko.tvmaniac.domain.watchlist.WatchlistTasksInitializer
-import dev.zacsweers.metro.AppScope
-import dev.zacsweers.metro.BindingContainer
-import dev.zacsweers.metro.ContributesTo
-import dev.zacsweers.metro.IntoSet
-import dev.zacsweers.metro.Provides
-
-@BindingContainer
-@ContributesTo(AppScope::class)
-public interface WatchlistTasksInitializerBindingContainer {
-    public companion object {
-        @Provides
-        @IntoSet
-        @AsyncInitializers
-        public fun provideWatchlistTasksInitializer(
-            bind: WatchlistTasksInitializer,
-        ): Initializer = Initializer { bind.init() }
-    }
-}
