--- domain/watchlist/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/watchlist/model/EpisodeBadge.kt
@@ -1,7 +1,0 @@
-package com.thomaskioko.tvmaniac.domain.watchlist.model
-
-public enum class EpisodeBadge {
-    PREMIERE,
-    NEW,
-    NONE,
-}
