+++ settings.gradle.kts
@@ -212,17 +212,17 @@ include(
     ":data:continue-watching:api",
     ":data:continue-watching:implementation",
     ":data:continue-watching:testing",
-    ":data:watchlist:api",
-    ":data:watchlist:implementation",
-    ":data:watchlist:testing",
+    ":data:watchlist-prefs:api",
+    ":data:watchlist-prefs:implementation",
+    ":data:watchlist-prefs:testing",
     ":data:watchproviders:api",
     ":data:watchproviders:implementation",
     ":data:watchproviders:testing",
     ":domain:calendar",
+    ":domain:continue-watching",
     ":domain:discover",
     ":domain:episode",
     ":domain:feature-flags",
-    ":domain:upnext",
     ":domain:followedshows",
     ":domain:genre",
     ":domain:library",
@@ -233,10 +233,10 @@ include(
     ":domain:seasondetails",
     ":domain:showdetails",
     ":domain:similarshows",
+    ":domain:sync-activity",
     ":domain:theme",
     ":domain:traktlists",
     ":domain:user",
-    ":domain:watchlist",
     ":domain:watchproviders",
     ":i18n:api",
     ":i18n:generator",
